requirements.txt
aiofiles==23.1.0
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.101.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
openai==1.4.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.0.0
requests==2.32.0
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.22.0
whisper==1.1.0
python-multipart==0.0.6

